src/main/java/LogServer.java
import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.List;

public class LogServer implements ILogServer {
    private MerkleTree merkleTree;

    public LogServer(String logFilePath) throws IOException {
        List<String> events = Files.readAllLines(Paths.get(logFilePath));
        this.merkleTree = new MerkleTree(events);
    }

    @Override
    public void appendEvent(String event) {
        // 逻辑：将事件添加到当前事件列表，然后重新构建Merkle树
        // 注意：这里可以优化，只更新树的相关部分，而不是完全重建
        merkleTree.appendEvent(event);
    }

    @Override
    public byte[] getCurrentRootHash() {
        return merkleTree.getRootHash();
    }

    @Override
    public List<byte[]> genPath(String event) {
        // 逻辑：生成并返回验证事件的审计路径
        return merkleTree.genPath(event);
    }

    @Override
    public List<byte[]> genProof(int oldTreeSize, int newTreeSize) {
        // 逻辑：生成并返回一致性证明
        return merkleTree.genProof(oldTreeSize, newTreeSize);
    }

    // Additional methods and logic to be implemented
}
